Add GNSS, signal quality, pressure and temperature

+++ LiveObjects/LiveObjects.h
@@ -2,7 +2,7 @@
 #define LIVEOBJECTS_H_INC
 
 LE_SHARED void liveobjects_connect(char* apikey, char* namespace, char* id, void* connectionHandler);
-LE_SHARED void liveobjects_pubData(char* streamid, char* payload, char* model, char* tags, int latitude, int longitude);
+LE_SHARED void liveobjects_pubData(char* streamid, char* payload, char* model, char* tags, double latitude, double longitude);
 LE_SHARED void liveobjects_pubCmdRes(char* jsonStr, int cid);
 LE_SHARED void liveobjects_pubConfig(char* key, char* type, char* value);
 LE_SHARED void liveobjects_pubConfigUpdateResponse(char* key, char* type, char* value, int cid);
